Fix for an exception that was thrown in browsers with localization support
The issue was:
Applications that provide multiple translations also have multiple AppMetas associated with them.
If the client uses a language that there is a translation for, e.g. "de_DE", two or four AppMetas will be loaded (en, en_US, de, de_DE) and will thus confuse the system.
The fix is to randomly select one of these AppMetas to be used for the rendering.
It would be better to have a fallback chain (TODO).